+++ code/DIY2/DIY2.ino
@@ -289,7 +289,7 @@ void loop() {
   // Otherwise, enter deep sleep.
   if (active) {
     if (sampling) {
-      double pressure = sensor.getPressure(ADC_4096);
+      float pressure = sensor.getPressure(ADC_4096);
       int temperature = sensor.getTemperature(CELSIUS, ADC_512); 
           
       logFile.printField(now.year(), '-');
